merging, setting up to run

+++ backend/src/main/kotlin/com/laba/it_planner/service/impl/ProjectServiceImpl.kt
@@ -9,11 +9,9 @@ import com.laba.it_planner.model.project.Project
 import com.laba.it_planner.model.project.repository.ProjectRepo
 import com.laba.it_planner.model.user.ProjectRole
 import com.laba.it_planner.repository.ProjectRepository
-import com.laba.it_planner.service.EmployeeService
-import com.laba.it_planner.service.OauthService
-import com.laba.it_planner.service.ProjectRepoService
-import com.laba.it_planner.service.ProjectService
+import com.laba.it_planner.service.*
 import org.springframework.stereotype.Service
+import java.security.Principal
 import java.time.LocalDateTime
 
 @Service
@@ -21,15 +19,17 @@ class ProjectServiceImpl(
     private val projectRepository: ProjectRepository,
     private val oauthService: OauthService,
     private val employeeService: EmployeeService,
-    private val projectRepoService: ProjectRepoService
+    private val projectRepoService: ProjectRepoService,
+    private val userInfoService: UserInfoService,
+    private val mailService: MailService
 ) : ProjectService {
 
     override fun createProject(
         projectCreateRequestDto: ProjectCreateRequestDto,
-        username: String
+        principal: Principal
     ): Project {
-        val user = oauthService.getByUsername(username)
-        if (projectRepository.findByNameAndByCreatedUser(projectCreateRequestDto.name, user.id).isPresent) {
+        val userInfo = userInfoService.getUserInfo(principal)
+        if (projectRepository.findByNameAndByCreatedUser(projectCreateRequestDto.name, userInfo.id).isPresent) {
             throw DataException(
                 "Project with name ${projectCreateRequestDto.name} already exists",
                 "PROJECT_CREATION_ERROR"
@@ -39,12 +39,12 @@ class ProjectServiceImpl(
                 Project(
                     name = projectCreateRequestDto.name,
                     creationDate = LocalDateTime.now(),
-                    createdUser = user
+                    createdUser = userInfo
                 )
             )
 
             val employee = Employee(
-                user = user,
+                user = userInfo,
                 project = result,
                 projectRole = ProjectRole.PROJECT_MANAGER
             )
@@ -70,7 +70,8 @@ class ProjectServiceImpl(
         val projects = getAllUsersProjects(username)
         val project = projectRepository.findById(projectId)
         if (project.isPresent && projects.contains(project.get())) {
-            val user = oauthService.getByUsername(employeeInviteDto.username)
+            val user = userInfoService.getUserInfo(employeeInviteDto.username)
+            mailService.sendInformationForm(user.email!!, "You have been invited to this project ${project.get().name}")
             return employeeService.createEmployee(
                 Employee(
                     user = user,
@@ -84,7 +85,12 @@ class ProjectServiceImpl(
     override fun deleteEmployee(projectId: Long, employeeId: Long, username: String) {
         val projects = getAllUsersProjects(username)
         val project = projectRepository.findById(projectId)
-        if (project.isPresent && projects.contains(project.get())) {
+        val employee = employeeService.getById(employeeId)
+        if (project.isPresent
+            && projects.contains(project.get())
+            && employee.user.username != username
+            && project.get().createdUser!!.username != username
+        ) {
             employeeService.deleteEmployee(employeeId)
         }
     }
@@ -94,14 +100,35 @@ class ProjectServiceImpl(
         employeeId: Long,
         employeeUpdateRoleDto: EmployeeUpdateRoleDto,
         name: String
-    ) : Employee{
+    ): Employee {
         val project = projectRepository.findById(projectId)
-        if(project.isPresent && project.get().createdUser?.username.equals(name)) {
-           return employeeService.changeRole(employeeId,employeeUpdateRoleDto.projectRole)
+        if (project.isPresent && project.get().createdUser?.username.equals(name)) {
+            return employeeService.changeRole(employeeId, employeeUpdateRoleDto.projectRole)
+        } else {
+            throw DataException(
+                "Project with id $projectId does not exist or it's not your's",
+                "PROJECT_NOT_FOUND_ERROR"
+            )
         }
-        else{
-            throw DataException("Project with id $projectId does not exist or it's not your's", "PROJECT_NOT_FOUND_ERROR")
+    }
+
+    override fun get(projectId: Long): Project {
+        return projectRepository.findById(projectId)
+            .orElseThrow { DataException("Project with id $projectId not found", "NOT_FOUND_ERROR") }
+    }
+
+    override fun getAllProjects(name: String): List<Project> {
+        return projectRepository.getAllUsersProjectsByUsername(name)
+    }
+
+    override fun deleteProject(projectId: Long, principal: Principal) {
+        if (employeeService.checkPermission(projectId, principal)) {
+            projectRepository.deleteById(projectId)
         }
+    }
+
+    override fun getByTaskId(taskId: Long): Project {
+        return projectRepository.findByTaskId(taskId)
     }
 
     override fun getAllUsersProjects(username: String): List<Project> {
